+++ crates/nix_rs/CHANGELOG.md
@@ -10,6 +10,8 @@
   - `nix_eval_attr_json`: No longer takes `default_if_missing`; instead (always) returns `None` if attribute is missing.
 - **`env::NixEnv`**
   - Clarify error message when `$USER` is not set
+- **``command`**
+  - Add `NixCmd::get()` to return flakes-enabled global command
 
 ## [0.5.0](https://github.com/juspay/nix-rs/compare/0.4.0...0.5.0) (2024-06-05)